pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "Small fastapi boilerplate"
authors = ["stupidcabbage <vanya@gutche.ru>"]
license = "MIT License"
packages = [{include = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
fastapi = "^0.115.6"
uvicorn = "^0.33.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
flake8 = "^7.1.1"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
wemake-python-styleguide = "^1.0.0"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
